requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
dj-database-url==2.2.0
Django==4.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.0
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
PyJWT==2.9.0
PyMySQL==1.1.1
pyopt==0.84
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
